dfa_test/gallery/serializer.py
from rest_framework import serializers

from gallery.models import Image


class ImageSerializer(serializers.ModelSerializer):
    class Meta:
        model = Image
        fields = "__all__"
        read_only_fields = ['owner']
